Add multilingual README support to enhance project accessibility
This commit introduces translated versions of the README, starting with Portuguese, to make the project more accessible and inclusive to contributors from different regions. Since this is an open-source project, ensuring that the instructions and goals are available in multiple languages is an important step towards increasing global participation. Additional languages will be added in the future to further extend the project's reach.